bluedroid: Add option for a device-specific bluedroid implementation
We are in the process of adding ANT+ sensor support for tiwlan-based
SEMC devices. These binary plugs expect a slightly different bluedroid
implementation which may be useful for others implementing ANT+ but
may cause problems for other devices.

This patch adds BOARD_CUSTOM_BLUEDROID which can be used to replace the
standard bluetooth.c with a custom implementation without breaking
things for devices that assume default functionality.

Change-Id: I5f5ba2b36c0ffe4ab83130122dc6bda9aeeb9fc6