Remove bt_enable log spam
A flood of "bt_enable: ret" -1" messages happens if bluetooth fails to
initialize properly. This is due to a message which seems redundant in
system/bluetooth/bluedroid/bluetooth.c

 LOGI("bt_enable: ret: %d, errno: %d", ret, errno);

If a failure happens after the 1000 retries a single more useful error
message is printing. The LOGI message above should be removed.

Change-Id: I8f7f3eca7eb7e0d7c2551a08ec976222ab95651a